Rename qFunction libCEED files with _qf.
